==> notebooks/01_data_preparation_zee.ipynb <==
import sys
repo_path = '/home/micael.verissimo/paper_lzt/exp-connection-hep/'
sys.path.insert(0, repo_path)


import os

import numpy as np
import pandas as pd

from src import build_pandas

from loguru import logger


def_dict = {f'cl_ring_{iring}' : f'cl_rings[{iring}]' for iring in range(100)}
def_dict.update({'mc_e_float'     : 'mc_e[0]',
                 'mc_et_float'    : 'mc_et[0]',
                 'mc_eta_float'   : 'mc_eta[0]',
                 'mc_phi_float'   : 'mc_phi[0]', 
                 'mc_pdgid_float' : 'mc_pdgid[0]'})
s_dict = def_dict.copy()
s_dict.update({'target': '1'})
b_dict = def_dict.copy()
b_dict.update({'target': '0'})

print(s_dict.keys(), b_dict.keys())

data_path = '~/cern_data/joao.pinto/simulation/v2.1.0/datasets'

zee  = os.path.join(data_path, 'user.joao.pinto.mc25_13TeV.250401.Pythia8EvtGen_Zee.etamax0.8.10k_stage_4.result.NTUPLE.merged.root')
jf17 = os.path.join(data_path, 'user.joao.pinto.mc25_13TeV.250401.Pythia8EvtGen_JF17.etamax0.8.10k_stage_4.result.NTUPLE.merged.root')

zee_mb = os.path.join(data_path, 'user.joao.pinto.mc25_13TeV.250401.Pythia8EvtGen_Zee.etamax0.8.avgmu100.sigmamu20.10k_stage_4.result.NTUPLE.merged.root')
jf17_mb = os.path.join(data_path, 'user.joao.pinto.mc25_13TeV.250401.Pythia8EvtGen_JF17.etamax0.8.avgmu100.sigmamu20.10k_stage_4.result.NTUPLE.merged.root')


zee_df = build_pandas([zee], 'physics', os.path.join(repo_path, 'data/processed'), 'zee.parquet', definitions=s_dict)
zee_mb_df = build_pandas([zee_mb], 'physics', os.path.join(repo_path, 'data/processed'), 'zee_mb.parquet', definitions=s_dict)

jf17_df = build_pandas([jf17], 'physics', os.path.join(repo_path, 'data/processed'), 'jf17.parquet', definitions=b_dict)
jf17_mb_df = build_pandas([jf17_mb], 'physics', os.path.join(repo_path, 'data/processed'), 'jf17_mb.parquet', definitions=b_dict)
