==> notebooks/06_crossvalidation.ipynb <==
import sys
repo_path = '/home/micael.verissimo/paper_lzt/exp-connection-hep/'
sys.path.insert(0, repo_path)

import os
import tqdm
import pickle
import collections
import itertools

import matplotlib as mpl
mpl.rcParams['text.usetex'] = True

import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import tensorflow as tf
tf.config.run_functions_eagerly(True)
tf.get_logger().setLevel('INFO')

from datetime import datetime
from sklearn.model_selection import StratifiedKFold
from loguru import logger

import logging
logger = logging.getLogger()
logger.setLevel(logging.CRITICAL)

from src.crossval import crossval_table

#
# Create the reference map
#
def create_op_dict(op, decoration='reference'):
    
    d = collections.OrderedDict( {
              # validation
              "max_sp_val"      : 'summary/max_sp_val',
              "max_sp_pd_val"   : 'summary/max_sp_pd_val#0',
              "max_sp_fa_val"   : 'summary/max_sp_fa_val#0',
              # Operation
              "max_sp_op"       : 'summary/max_sp_op',
              "max_sp_pd_op"    : 'summary/max_sp_pd_op#0',
              "max_sp_fa_op"    : 'summary/max_sp_fa_op#0',
              
              # op
              'pd_ref'    : decoration+"/"+op+"/pd_ref#0",
              #'fa_ref'    : decoration+"/"+op+"/fa_ref#0",
              #'sp_ref'    : decoration+"/"+op+"/sp_ref",
              'pd_val'    : decoration+"/"+op+"/pd_val#0",
              'fa_val'    : decoration+"/"+op+"/fa_val#0",
              'thr_val'   : decoration+"/"+op+"/threshold_val",
              'sp_val'    : decoration+"/"+op+"/sp_val",
              'pd_op'     : decoration+"/"+op+"/pd_op#0",
              'fa_op'     : decoration+"/"+op+"/fa_op#0",
              'sp_op'     : decoration+"/"+op+"/sp_op",
              'thr_op'    : decoration+"/"+op+"/threshold_op",

              # Counts
              'pd_ref_passed'    : decoration+"/"+op+"/pd_ref#1",
              #'fa_ref_passed'    : decoration+"/"+op+"/fa_ref#1",
              'pd_ref_total'     : decoration+"/"+op+"/pd_ref#2",
              #'fa_ref_total'     : decoration+"/"+op+"/fa_ref#2",
              'pd_val_passed'    : decoration+"/"+op+"/pd_val#1",
              'fa_val_passed'    : decoration+"/"+op+"/fa_val#1",
              'pd_val_total'     : decoration+"/"+op+"/pd_val#2",
              'fa_val_total'     : decoration+"/"+op+"/fa_val#2",
              'pd_op_passed'     : decoration+"/"+op+"/pd_op#1",
              'fa_op_passed'     : decoration+"/"+op+"/fa_op#1",
              'pd_op_total'      : decoration+"/"+op+"/pd_op#2",
              'fa_op_total'      : decoration+"/"+op+"/fa_op#2",
    })
    return d


op_names = ['tight', 'medium', 'loose']

tuned_info = collections.OrderedDict({})
for op in op_names:
    tuned_info[op] = create_op_dict(op, "reference")

#
# Et/eta bins edges
# 
etbins = [15, 30, 50, 100000]
etabins = [0.0, 0.8, 1.37, 1.54, 2.37, 2.50]
tuning_path = os.path.join(repo_path, 'data/models')

cv_ss    = crossval_table( tuned_info, etbins = etbins, etabins = etabins )
cv_ss.fill( os.path.join(tuning_path, 'ss_models/*.pic'), 'MLP-5-SS')

cv_v0    = crossval_table( tuned_info, etbins = etbins, etabins = etabins )
cv_v0.fill( os.path.join(tuning_path, 'rings_v0_models/*.pic'), 'MLP-5-rings_v0')

cv_v1    = crossval_table( tuned_info, etbins = etbins, etabins = etabins )
cv_v1.fill( os.path.join(tuning_path, 'rings_v1_models/*.pic'), 'MLP-5-rings_v1')

# obtain the best inits and best sorts and then 
best_inits_ss = cv_ss.filter_inits("max_sp_val")
best_sorts_ss = cv_ss.filter_sorts( best_inits_ss , 'max_sp_op')

best_inits_v0 = cv_v0.filter_inits("max_sp_val")
best_sorts_v0 = cv_v0.filter_sorts( best_inits_v0 , 'max_sp_op')

best_inits_v1 = cv_v1.filter_inits("max_sp_val")
best_sorts_v1 = cv_v1.filter_sorts( best_inits_v1 , 'max_sp_op')

best_inits = pd.concat([
                        best_inits_v0,
                        best_inits_v1,
                        best_inits_ss
                       ])

best_sorts = pd.concat([
                        best_sorts_v0,
                        best_sorts_v1,
                        best_sorts_ss
                       ])

print(len(best_inits))#(15bins x 5folds x 3op x 3inputs = 675)
print(len(best_sorts))#(15bins x 1folds x 3op x 3inputs = 135)

best_inits.to_csv(os.path.join(repo_path, 'data/processed/best_inits.csv'), index=False)
best_sorts.to_csv(os.path.join(repo_path, 'data/processed/best_sorts.csv'), index=False)

# verificar os tunings separando-os em regiões de low e high energy
cv    = crossval_table( tuned_info, etbins = etbins, etabins = etabins )
for iop in ['tight', 'medium', 'loose']:
    cv.dump_beamer_table( best_inits = best_inits.loc[(best_inits.op_name == iop) & (best_inits.et_bin <= 2)] ,
                          output = f'./cv_{iop}',
                          tags  = ['MLP-5-SS','MLP-5-rings_v0', 'MLP-5-rings_v1'],
                          title = iop+' Tunings LZT', toPDF=False
                        )
