==> notebooks/07_add_models_decision.ipynb <==
import sys
repo_path = '/home/micael.verissimo/paper_lzt/exp-connection-hep/'
sys.path.insert(0, repo_path)

import os
import json
import time
import pickle
import itertools

import numpy as np
import pandas as pd

from loguru import logger
from typing import List, Dict, Union
from src.constants import GeV, et_bins, eta_bins

import tensorflow as tf
from tensorflow.keras.models import Model
model_from_json = tf.keras.models.model_from_json

# Configurações do TensorFlow para evitar warnings
tf.config.run_functions_eagerly(True)
tf.data.experimental.enable_debug_mode()

# Configurações adicionais para melhor performance e menos warnings
tf.get_logger().setLevel('ERROR')  # Reduz logs verbosos
os.environ['TF_CPP_MIN_LOG_LEVEL'] = '2'  # Suprime warnings de compilação

# Tentar importar pandarallel, instalar se necessário
#from pandarallel import pandarallel
#pandarallel.initialize(progress_bar=True, nb_workers=6)

from concurrent.futures import ThreadPoolExecutor, ProcessPoolExecutor
import multiprocessing as mp

prefix = "cl_ring_%i"

# rings presample 
presample = [prefix % iring for iring in range(8 // 2)]

# EM1 list
sum_rings = 8
em1 = [prefix % iring for iring in range(sum_rings, sum_rings + (64 // 2))]

# EM2 list
sum_rings = 8 + 64
em2 = [prefix % iring for iring in range(sum_rings, sum_rings + (8 // 2))]

# EM3 list
sum_rings = 8 + 64 + 8
em3 = [prefix % iring for iring in range(sum_rings, sum_rings + (8 // 2))]

# HAD1 list
sum_rings = 8 + 64 + 8 + 8
had1 = [prefix % iring for iring in range(sum_rings, sum_rings + (4 // 2))]

# HAD2 list
sum_rings = 8 + 64 + 8 + 8 + 4
had2 = [prefix % iring for iring in range(sum_rings, sum_rings + (4 // 2))]

# HAD3 list
sum_rings = 8 + 64 + 8 + 8 + 4 + 4
had3 = [prefix % iring for iring in range(sum_rings, sum_rings + (4 // 2))]

input_cols_map = {
    'MLP-5-rings_v0' : [f"cl_ring_{idx}" for idx in range(100)],
    'MLP-5-rings_v1' : presample + em1 + em2 + em3 + had1 + had2 + had3,
    'MLP-5-SS' : ['cl_reta', 'cl_eratio', 'cl_rhad']    
}

sgn_df = pd.concat([pd.read_parquet(os.path.join(repo_path, f'data/processed/{iname}')) for iname in ['zee_avg300_50k.parquet']], axis=0)
bkg_df = pd.concat([pd.read_parquet(os.path.join(repo_path, f'data/processed/{iname}')) for iname in ['jf17_avg300_50k.parquet']], axis=0)

t_df = pd.concat([sgn_df, bkg_df], axis=0)
t_df = t_df.loc[t_df.cl_eta.abs() <= 2.5]
t_df = t_df.loc[t_df.cl_et >= 15000]

t_df.duplicated().sum()


tot = 0
for iet, (l_iet, h_iet) in enumerate(et_bins):
    for ieta, (l_ieta, h_ieta) in enumerate(eta_bins):
        print(f'Processing et bin = [{l_iet}, {h_iet}[ and eta bin = [{l_ieta}, {h_ieta}[')
        t_df.loc[(t_df.cl_et >= l_iet*GeV) & (t_df.cl_et < h_iet*GeV), 'et_bin'] = int(iet)
        t_df.loc[(t_df.cl_eta.abs() >= l_ieta) & (t_df.cl_eta.abs() < h_ieta), 'eta_bin'] = int(ieta)
        l_samples = len(t_df.loc[(t_df.et_bin == iet) & (t_df.eta_bin == ieta)])
        tot += l_samples

best_sorts = pd.read_csv(os.path.join(repo_path, 'data/processed/best_sorts.csv'))
best_sorts.head()

def get_models_from_path(model_path: str):
    """
    Get the models from the given path.
    
    Parameters:
    model_path (str): The path to the model.
    
    Returns:
    list: A list of models.
    """
    with open(model_path, 'rb') as f:
        tuned_file = pickle.load(f)
    
    model = tuned_file['model']
    weights = tuned_file['weights']

    model = model_from_json( json.dumps(model, separators=(',', ':')) ) #custom_objects={'RpLayer':RpLayer} )
    model.set_weights( weights )
    new_model = Model(model.inputs, model.layers[-2].output)
    return model, new_model

def generate_model_dict(train_tag: str) -> Dict[str, Dict[str, Union[Model, float]]]:
    """
    Add classifier information to the dataframe.
    
    Parameters:
    df (dataframe): The input dataframe.
    train_tag (str): The tag of the training.
    
    Returns:
    dataframe: The dataframe with classifier information added.
    """
    model_dict = {}
    model_dict['input_cols'] = input_cols_map[train_tag]
    models_df = best_sorts.loc[best_sorts.train_tag == train_tag].copy()
    for iet, ieta in itertools.product(range(len(et_bins)),
                                       range(len(eta_bins))):
        logger.info(f'Processing et bin = {iet} and eta bin = {ieta}')
        bin_key    = f'et{iet}_eta{ieta}'
        bin_models = models_df.loc[(models_df.et_bin == iet) & (models_df.eta_bin == ieta)]
        model_dict[bin_key] = {}

        sigmoid_model, linear_model = get_models_from_path(bin_models.file_name.unique()[0])
        model_dict[bin_key]['s_model'] = sigmoid_model
        model_dict[bin_key]['l_model'] = linear_model
        for iop in ['tight', 'medium', 'loose']:
            model_dict[bin_key][iop] = bin_models.loc[bin_models.op_name == iop].iloc[0].thr_val

    return model_dict
        

ss_models = generate_model_dict('MLP-5-SS')
v0_models = generate_model_dict('MLP-5-rings_v0')
v1_models = generate_model_dict('MLP-5-rings_v1')

def norm1(data):
    norms = np.abs(data.sum(axis=1))
    norms[norms == 0] = 1
    return data / norms[:, None]

def add_classifier_information(row: pd.Series, 
                               model_dict: Dict[str, Dict[str, Union[Model, float]]],
                               ):
    """
    Add classifier information to the dataframe.
    
    Parameters:
    df (dataframe): The input dataframe.
    model_dict (dict): The dictionary containing the models and thresholds.
    train_tag (str): The tag of the training.
    
    Returns:
    dataframe: The dataframe with classifier information added.
    """
    et_bin = int(row['et_bin'])
    eta_bin = int(row['eta_bin'])
    bin_key = f'et{et_bin}_eta{eta_bin}'
    
    input_cols = model_dict['input_cols']
    if 'rings' in model_dict['input_cols'][0]:
        # For rings models, we need to normalize the input
        inputs = row[input_cols].values.reshape(1, -1)
        inputs = norm1(inputs).astype(np.float32)
    else:
        # For SS models, we do not need to normalize the input 
        inputs = row[input_cols].values.reshape(1, -1).astype(np.float32)
    # Use the processed inputs for both models
    s_o = model_dict[bin_key]['s_model'].predict(inputs, verbose=0)[0][0]
    l_o = model_dict[bin_key]['l_model'].predict(inputs, verbose=0)[0][0]
    
    tight_passed = int(s_o >= model_dict[bin_key]['tight'])
    medium_passed = int(s_o >= model_dict[bin_key]['medium'])
    loose_passed = int(s_o >= model_dict[bin_key]['loose'])

    return s_o, l_o, tight_passed, medium_passed, loose_passed



# Função otimizada para processamento em lote
def process_batch_vectorized(df_batch, model_dict, train_tag):
    """
    Processa um lote de dados de forma vetorizada
    """
    results = []
    input_cols = model_dict['input_cols']
    df_batch[f'{train_tag}_sigmoid']       = 0
    df_batch[f'{train_tag}_linear']        = 0
    df_batch[f'{train_tag}_tight_passed']  = 0
    df_batch[f'{train_tag}_medium_passed'] = 0
    df_batch[f'{train_tag}_loose_passed']  = 0
    # Agrupar por et_bin e eta_bin para processar em lotes maiores
    for iet, ieta in itertools.product(range(len(et_bins)), range(len(eta_bins))):
        logger.info(f'Processing et bin = {iet} and eta bin = {ieta}')
        # Filtrar o DataFrame para o bin atual       
        bin_filter = (df_batch.et_bin == iet) & (df_batch.eta_bin == ieta) 
        l_df = df_batch.loc[bin_filter]
        bin_key = f'et{int(iet)}_eta{int(ieta)}'
        
        if bin_key not in model_dict:
            continue
            
        # Preparar dados em lote
        if 'rings' in input_cols[0]:
            # Para rings models, normalizar
            batch_inputs = l_df[input_cols].values
            batch_inputs = norm1(batch_inputs).astype(np.float32)
        else:
            # Para SS models, não normalizar
            batch_inputs = l_df[input_cols].values.astype(np.float32)
        
        # Predição em lote (muito mais eficiente)
        s_outputs = model_dict[bin_key]['s_model'].predict(batch_inputs, verbose=0).flatten()
        l_outputs = model_dict[bin_key]['l_model'].predict(batch_inputs, verbose=0).flatten()
        df_batch.loc[bin_filter, f'{train_tag}_sigmoid'] = s_outputs
        df_batch.loc[bin_filter, f'{train_tag}_linear']  = l_outputs
        
        # Aplicar thresholds vetorizadamente
        df_batch.loc[bin_filter, f'{train_tag}_tight_passed']  = (s_outputs >= model_dict[bin_key]['tight']).astype(int)
        df_batch.loc[bin_filter, f'{train_tag}_medium_passed'] = (s_outputs >= model_dict[bin_key]['medium']).astype(int)
        df_batch.loc[bin_filter, f'{train_tag}_loose_passed']  = (s_outputs >= model_dict[bin_key]['loose']).astype(int)
    
    return df_batch
# Aplicar classificadores de forma otimizada
map_dict = {
    'ss' : ss_models,
    'v0' : v0_models,
    'v1' : v1_models,
}

print(f"Processando {len(t_df)} amostras...")

for train_tag, classifiers in map_dict.items():
    logger.info(f'Processing train tag = {train_tag} com processamento vetorizado...')
    start_time = time.time()
    # Método 1: Processamento totalmente vetorizado (mais rápido)
    try:
        results = process_batch_vectorized(t_df, classifiers, train_tag)
            
    except Exception as e:
        logger.error(f"Erro no processamento vetorizado: {e}")
    
    end_time = time.time()
    logger.info(f'Completed {train_tag} in {end_time - start_time:.2f} seconds')
logger.info("Processamento concluído!")

t_df.head()

t_df.duplicated().sum()

t_df.to_parquet(os.path.join(repo_path, 'data/processed/lzt_zee_jf17_avg300_50k_classified.parquet'))
