==> notebooks/09_make_plots.ipynb <==
import sys
repo_path = '/home/micael.verissimo/paper_lzt/exp-connection-hep/'
sys.path.insert(0, repo_path)

import os
import json
import time
import pickle
import itertools

import numpy as np
import pandas as pd

from loguru import logger
from typing import List, Dict, Union
from src.constants import GeV, et_bins, eta_bins

import matplotlib.pyplot as plt 
import tensorflow as tf
from tensorflow.keras.models import Model
model_from_json = tf.keras.models.model_from_json

# Configurações do TensorFlow para evitar warnings
tf.config.run_functions_eagerly(True)
tf.data.experimental.enable_debug_mode()

# Configurações adicionais para melhor performance e menos warnings
tf.get_logger().setLevel('ERROR')  # Reduz logs verbosos
os.environ['TF_CPP_MIN_LOG_LEVEL'] = '2'  # Suprime warnings de compilação

def plot_stats(n_samples, save_path, output_name):
    '''
    '''
    etbins  = [f'[{l_et}, {h_et}[ GeV' for l_et, h_et in et_bins]
    etabins = [f'[{l_eta}, {h_eta}[' for l_eta, h_eta in eta_bins]
    # plot the samples
    fig, ax = plt.subplots(figsize=(15, 20))
    # Using matshow here just because it sets the ticks up nicely. imshow is faster.
    data = np.zeros((len(etbins), len(etabins)))
    ax.matshow(data, cmap='Greys')
    # anottate the samples
    for (i, j), z in np.ndenumerate(data):
        s, b = n_samples['et%i_eta%i' %(i,j)]
        s_b = (float(s)/float(b))*100
        ax.text(j, i-0.05, 'signal: %i' %s, ha='center', va='center', color='blue', fontsize=12)
        ax.text(j, i+0.05, 'background: %i' %b, ha='center', va='center', color='red', fontsize=12)
        ax.text(j, i+0.15, 's/b (%%): %1.2f' %(s_b), ha='center', va='center', color='black', fontsize=12)

    plt.axhline(y=1.5, lw=1.5, color='black')
    plt.axhline(y=0.5, lw=1.5, color='black')

    for ix in [0.5, 1.5, 2.5, 3.5]:
        plt.axvline(x=ix, lw=1.5, color='black')
    # change the bin labels
    ax.set_ylabel(r'$E_T$')
    ylabels = [''] + etbins + ['']
    ax.set_yticklabels(ylabels)

    ax.set_xlabel(r'$\eta$')
    xlabels = [''] + etabins + ['']
    ax.set_xticklabels(xlabels)

    plt.savefig(os.path.join(save_path, f'{output_name}.png'), dpi=200, facecolor='white', bbox_inches='tight')

trn_samples = {}
tst_samples = {}
for iet, (l_iet, h_iet) in enumerate(et_bins):
    for ieta, (l_ieta, h_ieta) in enumerate(eta_bins):
        print(f'Processing et bin = [{l_iet}, {h_iet}[ and eta bin = [{l_ieta}, {h_ieta}[')
        m_df.loc[(m_df.cl_et >= l_iet*GeV) & (m_df.cl_et < h_iet*GeV), 'et_bin'] = int(iet)
        m_df.loc[(m_df.cl_eta.abs() >= l_ieta) & (m_df.cl_eta.abs() < h_ieta), 'eta_bin'] = int(ieta)
        trn_samples[f'et{iet}_eta{ieta}'] = (len(m_df.loc[(m_df.et_bin == iet) & (m_df.eta_bin == ieta) & (m_df.target == 1)]),
                                             len(m_df.loc[(m_df.et_bin == iet) & (m_df.eta_bin == ieta) & (m_df.target != 1)]))
        
        t_df.loc[(t_df.cl_et >= l_iet*GeV) & (t_df.cl_et < h_iet*GeV), 'et_bin'] = int(iet)
        t_df.loc[(t_df.cl_eta.abs() >= l_ieta) & (t_df.cl_eta.abs() < h_ieta), 'eta_bin'] = int(ieta)
        tst_samples[f'et{iet}_eta{ieta}'] = (len(t_df.loc[(t_df.et_bin == iet) & (t_df.eta_bin == ieta) & (t_df.target == 1)]),
                                             len(t_df.loc[(t_df.et_bin == iet) & (t_df.eta_bin == ieta) & (t_df.target != 1)]))

m_df.cl_eratio.hist(bins=100)

m_df.cl_reta.describe()

m_df.cl_rhad.describe()

m_df.cl_rhad.hist(bins=100)

plot_stats(trn_samples, os.path.join(repo_path, 'data/processed'), 'train_samples')
plot_stats(tst_samples, os.path.join(repo_path, 'data/processed'), 'test_samples')

t_df = pd.read_parquet(os.path.join(repo_path, 'data/processed/lzt_zee_jf17_avg300_50k_classified.parquet'))

t_df.head()


for iet, ieta in itertools.product(range(3), range(5)):
    print(f'Processing et bin = {iet} and eta bin = {ieta}')
    aux_df = t_df.loc[(t_df.et_bin == iet) & (t_df.eta_bin == ieta) & (t_df.target == 1)]
    tot = len(aux_df)
    for imodel in ['ss', 'v0', 'v1']:
        fig, ax = plt.subplots(figsize=(10, 6))
        ax.hist(aux_df[f'{imodel}_sigmoid'], bins=np.arange(0, 1.01, step=.01), label=f'{imodel} output', alpha=0.5)
        plt.savefig(os.path.join(repo_path, 'data/reports', 'figures', f'hist_{imodel}_et{iet}_eta{ieta}.png'), dpi=200, facecolor='white', bbox_inches='tight')
        plt.close()
        for iop in ['tight', 'medium', 'loose']:
            
            passed = len(aux_df.loc[aux_df[f'{imodel}_{iop}_passed'] == 1])
            print(f'  Model: {imodel}, OP: {iop}, Passed: {passed} ({(passed/tot)*100:.2f}%)')
        print(10*'~~')
    print(10*'-.-')
    print('\n')
    

et, eta = 2, 0

print(list(t_df.columns))
